Refactor: 폴더 구조 변경 및 안쓰는 파일 삭제
