+++ src/main/java/com/kbulkup/counseling/dto/response/CounselingCreateResponseDTO.java
@@ -1,6 +1,5 @@
 package com.kbulkup.counseling.dto.response;
 
-import com.kbulkup.counseling.domain.Counseling;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
