--- src/main/java/com/kbulkup/counseling/mapper/CounselingMapper.java
@@ -1,27 +1,0 @@
-package com.kbulkup.counseling.mapper;
-
-import com.kbulkup.counseling.domain.Counseling;
-import com.kbulkup.counseling.dto.response.CounselingListResponseDTO;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Mapper
-public interface CounselingMapper {
-
-    List<CounselingListResponseDTO> findByUserId(Long userId);
-
-    Counseling findByTraineeAndTrainer(@Param("traineeId") Long traineeId, @Param("trainerId") Long trainerId);
-
-    Counseling findByRoomId(@Param("roomId") String roomId);
-
-    Boolean existsByUserIdAndTrainingId(@Param("userId") Long userId,  @Param("trainingId") Long trainingId);
-
-    void insertCounseling(Counseling counseling);
-
-    void updateLatestMessage(@Param("roomId") String roomId, @Param("latestMessage") String latestMessage, @Param("latestAt") LocalDateTime latestAt);
-
-    void updateStatusToExpired(@Param("roomId") String roomId, @Param("status") String status);
-}
