--- src/main/resources/mappers/counseling/CounselingMapper.xml
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.kbulkup.counseling.mapper.CounselingMapper">
-
-    <select id="findByUserId"
-            parameterType="long"
-            resultType="com.kbulkup.counseling.dto.response.CounselingListResponseDTO">
-            SELECT
-                c.room_id,
-                c.status,
-                c.latest_message,
-                c.latest_at,
-                c.expires_at,
-                t.title AS training_title,
-                CASE WHEN #{userId} = c.user_id THEN trainer.username ELSE trainee.username END AS opponent_name,
-                CASE WHEN #{userId} = c.user_id THEN trainer.user_profile_url ELSE trainee.user_profile_url END AS opponent_profile_url
-            FROM counselings c
-            JOIN users trainee ON trainee.user_id = c.user_id
-            JOIN users trainer ON trainer.user_id = c.trainer_id
-            JOIN trainings t ON t.training_id = c.training_id
-
-            WHERE #{userId} IN (c.user_id, c.trainer_id)
-            ORDER BY c.latest_at DESC
-    </select>
-
-    <select id="findByTraineeAndTrainer" resultType="com.kbulkup.counseling.domain.Counseling">
-        SELECT *
-        FROM counselings
-        WHERE user_id = #{traineeId}
-          AND trainer_id = #{trainerId}
-        LIMIT 1
-    </select>
-
-    <select id="findByRoomId" resultType="com.kbulkup.counseling.domain.Counseling">
-        SELECT *
-        FROM counselings
-        WHERE room_id = #{roomId}
-        LIMIT 1
-    </select>
-
-    <select id="existsByUserIdAndTrainingId" resultType="java.lang.Boolean">
-        SELECT EXISTS (
-            SELECT 1
-            FROM counselings
-            WHERE user_id = #{userId}
-              AND training_id = #{trainingId}
-        )
-    </select>
-
-    <insert id="insertCounseling" parameterType="com.kbulkup.counseling.domain.Counseling">
-        INSERT INTO counselings (
-            user_id,
-            trainer_id,
-            training_id,
-            room_id,
-            status,
-            latest_message,
-            latest_at,
-            start_at,
-            expires_at
-        ) VALUES (
-            #{userId},
-            #{trainerId},
-            #{trainingId},
-            #{roomId},
-            #{status},
-            #{latestMessage},
-            #{latestAt},
-            #{startAt},
-            #{expiresAt}
-        )
-    </insert>
-
-    <update id="updateLatestMessage">
-        UPDATE counselings
-        SET latest_message = #{latestMessage},
-            latest_at = #{latestAt}
-        WHERE room_id = #{roomId}
-    </update>
-
-    <update id="updateStatusToExpired">
-        UPDATE counselings
-        SET status = #{status}
-        WHERE room_id = #{roomId}
-    </update>
-</mapper>
